% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfDialectDisFile}
\alias{dfDialectDisFile}
\title{Definition of Format of Disease Model File}
\format{A data frame with 4 variables:
\describe{
  \item{category}{XXXX}
  \item{parameter}{parameter name}
  \item{required}{whether the parameter is required or optional}
  \item{type}{parameter's value type, e.g. character, integer, proportion, positive, characterVector, table}
  ...
}}
\usage{
dfDialectDisFile
}
\description{
Definition of Format of Disease Model File.
Essentially a list of parameters, whether they are required or optional, and what values they can take.
}
\keyword{datasets}
