% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfPed3Gen1Aff}
\alias{dfPed3Gen1Aff}
\title{3 Generation Pedigree.}
\format{A data frame with variables corresponding to ped file format:
\describe{
  \item{famid}{family id}
  \item{id}{person id}
  \item{fatherid}{father id}
  \item{motherid}{mother id}
  \item{sex}{sex}
  \item{affected}{affected status, 1=affected, 0=unaffected, Na=unknown}
  \item{age}{age}
  \item{deceased}{1=deceased,0=living}
  ...
}}
\usage{
dfPed3Gen1Aff
}
\description{
3 Generation Pedigree.
}
\keyword{datasets}
