% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisFileReaderSrc.R
\name{fnCalcDiseaseModel}
\alias{fnCalcDiseaseModel}
\title{Calculates disease model based on epidemiological findings}
\usage{
fnCalcDiseaseModel(lDis)
}
\arguments{
\item{lDis}{list containing disease epidemiological findings}
}
\value{
lDis list containing disease epidemiological findings and disease model parameters
}
\description{
Calculates liability threshold based disease model based on epidemiological findings
}
