% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnCalcExprPropLifetimeRisk}
\alias{fnCalcExprPropLifetimeRisk}
\title{Calculates expressed proportion of lifetime risk for pedigree members}
\usage{
fnCalcExprPropLifetimeRisk(lDis, dfPed)
}
\arguments{
\item{lDis}{disease model}

\item{dfPed}{pedigree}
}
\value{
pedigree with expressedProportionOfLifetimeRisk
}
\description{
If disease model congenital then EPLR=1 for all. Otherwise
Attaches an age of onset curve to each person according to their personal attributes.
Interpolates each persons EPLR from their age and age of onset curve
}
