% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnCalcPedCov}
\alias{fnCalcPedCov}
\title{Calc pedigree covariance matrix for trait risk factors, liability and its partitions}
\usage{
fnCalcPedCov(mCovPerson, dfPed)
}
\arguments{
\item{mCovPerson}{XXXX}

\item{dfPed}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
