% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnCalcPedDistributionPosteriorToObsRiskFactors}
\alias{fnCalcPedDistributionPosteriorToObsRiskFactors}
\title{XXXX}
\usage{
fnCalcPedDistributionPosteriorToObsRiskFactors(vMeanAllPed, mCovAllPed,
  vObsRFValues)
}
\arguments{
\item{vMeanAllPed}{XXXX}

\item{mCovAllPed}{XXXX}

\item{vObsRFValues}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
