% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnCollateObservedRiskFactors}
\alias{fnCollateObservedRiskFactors}
\title{Collate observations on risk factors}
\usage{
fnCollateObservedRiskFactors(lDis, dfPed)
}
\arguments{
\item{lDis}{XXXX}

\item{dfPed}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
