% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedigreeSrc.R
\name{fnConstructPedigree}
\alias{fnConstructPedigree}
\title{Constructs a kinship2::pedigree object}
\usage{
fnConstructPedigree(dfPed)
}
\arguments{
\item{dfPed}{ped format style data.frame}
}
\value{
kinship2:pedigree object
}
\description{
Constructs a kinship2::pedigree object from ped format style data.frame.
Assumes a valid pedigree, i.e. pedigree checked via fnValidatePedigree().
Collates twin info.
}
