% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnConvertCovArrayIntoMatrix}
\alias{fnConvertCovArrayIntoMatrix}
\title{Create population cov matrix for a single person}
\usage{
fnConvertCovArrayIntoMatrix(aCov)
}
\arguments{
\item{aCov}{XXXX}
}
\value{
mCovPerson XXXX
}
\description{
XXXX
}
