% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnEstimateRiskFromTotLiaSample}
\alias{fnEstimateRiskFromTotLiaSample}
\title{Calculates risks based on a sample liability dataset}
\usage{
fnEstimateRiskFromTotLiaSample(dfPed, mTotLiaSample)
}
\arguments{
\item{dfPed}{XXXX}

\item{mTotLiaSample}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
