% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnHeatMapCovMatrixSrc.R
\name{fnHeatMapCovMatrix}
\alias{fnHeatMapCovMatrix}
\title{Plots a heatmap representation of a covariance matrix}
\usage{
fnHeatMapCovMatrix(mat_data, nofSignifDigits = 2)
}
\arguments{
\item{mat_data}{matrix to be plotted}

\item{nofSignifDigits}{number of significant digits for cell contents}
}
\description{
Plots a heatmap representation of a covariance matrix
}
