% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnPrepareForRiskPrediction}
\alias{fnPrepareForRiskPrediction}
\title{XXXX}
\usage{
fnPrepareForRiskPrediction(dfPed, lDis, bVerbose = F)
}
\arguments{
\item{dfPed}{pedigree}

\item{lDis}{disease model}

\item{bVerbose}{boolean for reporting plots}
}
\value{
list containing liability draws, risk estimates, etc
}
\description{
XXXX
}
