% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisFileReaderSrc.R
\name{fnReadDisFromFile}
\alias{fnReadDisFromFile}
\title{Reads a disease model file into a named list of values
XXXX}
\usage{
fnReadDisFromFile(sFileDis)
}
\arguments{
\item{sFileDis}{disease model file}
}
\value{
lDis disease model
}
\description{
Reads a disease model file into a named list of values
XXXX
}
