% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileReaderSrc.R
\name{fnReadFile}
\alias{fnReadFile}
\title{XXXX}
\usage{
fnReadFile(sFile, vsTokenNameValuePair, vsTokenLinesAfter, vsTokenTableAfter)
}
\arguments{
\item{sFile}{XXXX}

\item{vsTokenNameValuePair}{XXXX}

\item{vsTokenLinesAfter}{XXXX}

\item{vsTokenTableAfter}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
