% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisFileReaderSrc.R
\name{fnReadFileDis}
\alias{fnReadFileDis}
\title{Reads a disease model from file}
\usage{
fnReadFileDis(sFileDis)
}
\arguments{
\item{sFileDis}{path to disease model file}
}
\value{
disease model
}
\description{
Parses the supplied disease model file according to file format defined in the supplied disease model file format definition.
}
