% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalRiskFactorSrc.R
\name{fnRunOptim}
\alias{fnRunOptim}
\title{Optimises liability mean and variance for a categorical risk factor}
\usage{
fnRunOptim(lifetimeRisk, dfRF, sMethod, lControl)
}
\arguments{
\item{lifetimeRisk}{population lifetime risk for the disease}

\item{dfRF}{directly calculable risk factor parameters}

\item{sMethod}{optimisation method}

\item{lControl}{optimisation method tuning parameters}
}
\value{
list containing solution
}
\description{
Uses optimisation to find the liability mean and variance for the strata of a categorical risk factor
}
