% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitSrc.R
\name{fnSS}
\alias{fnSS}
\title{Calculates descriptive stats on data.frame columns}
\usage{
fnSS(dfData)
}
\arguments{
\item{dfData}{data-table to be described}
}
\value{
data.frame containing descriptive stats
}
\description{
Calculates descriptive stats on data.frame columns
}
