% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalRiskFactorSrc.R
\name{fnValidateCategoricalRiskFactorInputs}
\alias{fnValidateCategoricalRiskFactorInputs}
\title{Validates categorical risk factor inputs, calculates directly computable parameters}
\usage{
fnValidateCategoricalRiskFactorInputs(lifetimeRisk, vFreq, vRelRisk)
}
\arguments{
\item{lifetimeRisk}{population lifetime risk for the disease}

\item{vFreq}{population frequency vector for strata (must sum to 1)}

\item{vRelRisk}{relative risk vector for strata (one element must equal 1, this identifies the reference stratum)}
}
\value{
data.frame containing directly calculable risk factor parameters
}
\description{
Validates categorical risk factor inputs are in correct ranges etc.
Calculates directly computable risk factor parameters.
Checks whether these are valid - some input combos are invalid, e.g. produce stratum risks > 1.
}
