% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedigreeSrc.R
\name{fnValidatePedigree}
\alias{fnValidatePedigree}
\title{Partial valdidation of a pedigree data.frame}
\usage{
fnValidatePedigree(dfPed)
}
\arguments{
\item{dfPed}{pedigree data.frame}
}
\value{
dfPed
}
\description{
Partial valdidation of a pedigree data.frame
}
