% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lDisDepression}
\alias{lDisDepression}
\title{Disease Model for Major Depression}
\format{A named list containing epidemiological findings defining the model:
\describe{
  \item{title}{title of the model}
  \item{description}{description of the model}
  \item{other}{etc.}
  ...
}}
\usage{
lDisDepression
}
\description{
Disease Model for Major Depression.
}
\keyword{datasets}
