% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalRiskFactorSrc.R
\name{plotRefStraMean}
\alias{plotRefStraMean}
\title{Generates plots indicative of whether Categorical risk factor optimisation converged}
\usage{
plotRefStraMean(sMethod, args, dfLog)
}
\arguments{
\item{sMethod}{optimisation method used}

\item{args}{categorical risk factor info}

\item{dfLog}{log containing history of optimisation}
}
\description{
Generates 2 plots x axis being number of iterations, y axis being
- categorical risk factor reference stratum mean
- discrepancy
The reference stratum mean should stabalise.
the discrepancy should fall to a very low value.
}
