% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateRiskSrc.R
\name{fnSampleConditionedLiaDistribution}
\alias{fnSampleConditionedLiaDistribution}
\title{XXXX}
\usage{
fnSampleConditionedLiaDistribution(dfPed, mPostCov, mPostMean,
  nofBurnIn = 100, nofIter = 9000)
}
\arguments{
\item{dfPed}{XXXX}

\item{mPostCov}{XXXX}

\item{mPostMean}{XXXX}

\item{nofBurnIn}{XXXX}

\item{nofIter}{XXXX}
}
\value{
XXXX
}
\description{
XXXX
}
